+++ pyremo/cmor/__init__.py
@@ -1,0 +1,5 @@
+from .remo_cmor import (
+    prepare_variable,
+    cmorize_variable,
+    to_cftime,
+)