+++ pyremo/core/cal.py
@@ -1,4 +1,4 @@
-"""this module handles absolute calendars (for a-files.)
+"""this module handles absolute calendars that occur in REMO datasets.
 """
 
 import datetime as dt
@@ -54,6 +54,8 @@ class AbsoluteCalendar:
         """convert a numeric absolute date value to a datetime object."""
         frac, whole = math.modf(num)
         date_str = str(int(whole))
+        if date_str[6:8] == "00":
+            date_str = date_str[0:6] + "15"
         # date = pd.to_datetime(date_str, format=self.fmt) #dt.datetime.strptime(date_str, self.fmt)
         date = dt.datetime.strptime(date_str[0:8], self.fmt)
         datetime = roundTime(
@@ -71,3 +73,35 @@ class AbsoluteCalendar:
                 datetime.minute,
             )
         return datetime
+
+
+def parse_dates(ds, use_cftime=False):
+    """Update the time axis of a REMO dataset.
+
+    Updates the time axis of a REMO dataset containing an absolute time axis.
+
+    Parameters
+    ----------
+    ds : xr.Dataset
+        Dataset with absolute time axis.
+    use_cftime: bool
+        Use cftime objects instead of datetime objects.
+    """
+    ds["time"] = parse_absolute_time(ds.time, use_cftime=use_cftime)
+    return ds
+
+
+def parse_absolute_time(time, use_cftime=False):
+    """Update a time axis containg fractional absolute dates.
+
+    Updates fractional absolute dates to relative dates.
+
+    Parameters
+    ----------
+    time : array like or xr.DataArray
+        Time axis containing absolute dates as float or int.
+    use_cftime: bool
+        Use cftime objects instead of datetime objects.
+    """
+    parser = AbsoluteCalendar()
+    return [parser.num2date(date, use_cftime) for date in time]