+++ tests/test_domain.py
@@ -4,12 +4,9 @@ import pytest
 
 import pyremo as pr
 
+
 def test_domain_dataset():
 
     eur11 = pr.remo_domain("EUR-11")
     eur22 = pr.remo_domain("EUR-22")
     eur44 = pr.remo_domain("EUR-44")
-
-
-
-