+++ docs/whats_new.rst
@@ -18,6 +18,7 @@ New Features
 - Included physics package with xarray API.
 - Included preprocessing interface for xarray data structures.
 - Included ERA5 cmorizer.
+- Includes experimental cmorization functions (:pull:`20`, :pull:`21`, :pull:`22`).
   
 Internal Changes
 ~~~~~~~~~~~~~~~~